rename: useSearchMember.ts -> useSearchSameCompanyMember.ts #313
