+++ frontend/src/components/ui/form/SearchMemberFieldFromProjectCreate.tsx
@@ -3,7 +3,7 @@ import { AVATAR_STYLE } from 'consts/avatarStyle'
 import { calculateMinSizeBasedOnFigma } from 'utils/calculateSizeBasedOnFigma'
 import { convertIntoRGBA } from 'utils/color/convertIntoRGBA'
 import styled, { css } from 'styled-components'
-import { useSearchMember } from 'hooks/useSearchMember'
+import { useSearchMember } from 'hooks/useSearchSameCompanyMember'
 import { useCalculateOverUsers } from 'hooks/useCalculateOverUsers'
 import { UserAvatarIcon } from 'components/ui/avatar/UserAvatarIcon'
 import { UserCount } from 'components/ui/avatar/UserCount'
